feat: 로그아웃 로직 수정 (#9)
