+++ src/pages/admin/_AdminPage.tsx
@@ -5,13 +5,10 @@ import {
 } from 'react-icons/pi';
 import { MenuContainer } from '@components/MenuContainer';
 import { Button } from '@components/Button';
-import { removeToken } from '@utils/handleAuth';
-import { useAuthStore } from '@stores/auth-stores';
-import { useNavigate } from 'react-router-dom';
+import useAuth from '@/hooks/useAuth';
 
 export const AdminPage = () => {
-  const { setUserInfo, setIsLoggedIn } = useAuthStore();
-  const navigate = useNavigate();
+  const { logout } = useAuth();
 
   const menuList = [
     {
@@ -22,23 +19,11 @@ export const AdminPage = () => {
     { id: 'news-upload', icon: <PiNotePencilLight />, menuName: '소식 업로드' },
     { id: 'form-upload', icon: <PiNotebookLight />, menuName: '폼 업로드' },
   ];
+
   return (
     <div className="flex flex-col gap-3">
       <MenuContainer menuList={menuList} />
-      <Button
-        content="로그아웃"
-        size="full"
-        onClick={() => {
-          removeToken();
-          setUserInfo({
-            name: '',
-            role: '',
-            studentID: '',
-          });
-          setIsLoggedIn(false);
-          navigate('/start');
-        }}
-      />
+      <Button content="로그아웃" size="full" onClick={logout} />
     </div>
   );
 };
